Rework the backend architecture to follow redirects.
This tackles issue #2: currently, if the targeturl of a proxy request
results in a redirect, that redirect gets relayed to the client. That's
appropriate for transparent HTTP proxies but not for our CORS-proxy
functionality.

As far as I can tell, there's no out-of-the-box solution to provide a
"wget-like" HTTP server that's nearly a passthrough, but follows
redirects. So we roll our own, in a simple Python/Tornado HTTP server.
Of course it's never as simple as you'd hope, but ...

We maintain the Squid frontend to cache things.